% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{gen_xs_corr}
\alias{gen_xs_corr}
\title{Generate correlated covariates X}
\usage{
gen_xs_corr(n, p, covar = 0.5)
}
\arguments{
\item{n}{Number of observations to generate}

\item{p}{Number of dimensions for each observation (includes the
intercept)}

\item{covar}{Off-diagonal element of covariance matrix.}
}
\value{
An n * (p-1) matrix of generated covariates
}
\description{
Generate the X_i independently, but with each component of X_i
distributed Multivariate Normal with mean 0 and covariance matrix
with 1 on the diagonal and \code{covar} on the off-diagonal. Exported
to the user.
}
\examples{
gen_xs_corr(10, 2)
gen_xs_corr(10, 2, 0.5)

set.seed(1)
# these should be equivalent
x1 <- gen_xs_corr(10, 2, 0)
set.seed(1)
x2 <- gen_xs_default(10, 2)
all(x1 == x2)
}

