% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{gen_xs_default}
\alias{gen_xs_default}
\title{Default function to generate covariates X.}
\usage{
gen_xs_default(n, p)
}
\arguments{
\item{n}{Number of observations to generate}

\item{p}{Number of dimensions for each observation (includes the
intercept)}
}
\value{
An n * (p-1) matrix of generated covariates
}
\description{
Generates the X_i independently, with each component of X_i
also being an independent Normal(0, 1). Exported to the user.
}
\examples{
gen_xs_default(10, 2)
}

