% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1proj.R
\docType{package}
\name{methods1proj}
\alias{methods1proj}
\alias{methods1proj-package}
\title{methods1proj: A package for running simulations for Methods I project}
\description{
methods1proj: A package for running simulations for Methods I project
}

