% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{pick_fit_model}
\alias{pick_fit_model}
\title{Pick the best linear regression model and fit it.}
\usage{
pick_fit_model(xs, ys, method = c("step", "leaps"), direction = "both")
}
\arguments{
\item{xs}{Covariates}

\item{ys}{Responses}

\item{method}{If leaps, run leaps. Otherwise, if step, run step-wise
selection}

\item{direction}{Only valid of \code{method = "step"}. Run step-wise
selection using \code{step} in the specified direction}
}
\value{
95% confidence intervals from the selected model.
}
\description{
Calls \code{pick_fit_model_leaps} or \code{pick_fit_model_step} to
actually do the underlying work. NOT exported to the user.
}

