% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{pick_fit_model_leaps}
\alias{pick_fit_model_leaps}
\title{Pick the best model using \code{leaps} and return its fitted
confidence intervals.}
\usage{
pick_fit_model_leaps(xs, ys, method)
}
\arguments{
\item{xs}{covariates}

\item{ys}{responses}

\item{method}{A valid argument to the \code{methods} argument of the
leaps command}
}
\value{
95% confidence intervals from the selected model.
}
\description{
Pick the best model using \code{leaps} and return its fitted
confidence intervals. NOT exported to the user.
}

