% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{plot_sim}
\alias{plot_sim}
\title{Draws coverage plots for one run of the simulation.}
\usage{
plot_sim(res_df, true_betas)
}
\arguments{
\item{res_df}{A data.frame returned by \link{\code{run_sim}}.}

\item{true_betas}{A vector of true values of beta, including the
intercept}
}
\value{
A list of ggplot2 objects, one per coefficient
}
\description{
Not really recommended because you ideally want to run the simulation
multiple times to get a sense of the variation. NOT exported to the
user
}

