% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{run_coverage_selected}
\alias{run_coverage_selected}
\title{Obtain coverage and selection probabilities}
\usage{
run_coverage_selected(nreps, n, true_betas, select_method = c("step",
  "leaps"), direction = "both", split_prop = -1,
  gen_dist = gen_xs_default, error_dist = rnorm, ...)
}
\arguments{
\item{nreps}{Number of repetitions to run, or how many datasets to
generate}

\item{n}{Number of observations per dataset.}

\item{true_betas}{A vector of true values of regression coefficients.
Should include the intercept as the first item, beta1 as the second,
etc.}

\item{select_method}{If leaps, run leaps. Otherwise, if step, run
step-wise selection}

\item{direction}{Only valid of \code{method = "step"}. Run step-wise
selection using \code{step} in the specified direction}

\item{split_prop}{If non-positive, run naive method, where the
confidence #' intervals are calculated on the same data used to
select the variables. Otherwise, if 0 < \code{split_prop} < 1,
randomly subset \code{split_prop} of the data to be in the training
set, and the other part to be in the test set. Run variable selection
on the training set, and report the final confidence interval as
calculated by fitting the selected model on the test set.}

\item{gen_dist}{A function specifiying how to generate the
covariates. Should take \code{n} as its first argument, corresponding
to the number of observations, and \code{p} as the second,
corresponding to the number of dimensions per observation (including
the intercept, so each X_i has p-1 dimensions).}

\item{error_dist}{A function specifying how to generate the errors.
Should take \code{n} as the first argument, specifying the number of
observations}

\item{...}{Additional arguments to gen_dist}
}
\value{
A list \code{lst} containing two vectors. The vector
\code{lst$coverages} is named, with each element corresponding to
the coverage probability of one coefficient in one iteration of the
simulation (is NA if that coefficient was never selected). The
vector \code{lst$select_probs} is like \code{lst$coverages}, but
instead contains the selection probabilities.
}
\description{
Calls \link{\code{run_sim}} once and processes the result to obtain
 coverage and selection probabilities.
}
\details{
Might be a bug with \code{get_coverage} and this function. Run with
n = 500 and split_prob = 0.8 and they're all NA.
}

