% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{run_sim}
\alias{run_sim}
\title{Run one replicate of the simulation.}
\usage{
run_sim(nreps, n, true_betas, select_method = c("step", "leaps"),
  direction = "both", split_prop = -1, gen_dist = gen_xs_default,
  error_dist = rnorm, ...)
}
\arguments{
\item{nreps}{Number of repetitions to run, or how many datasets to
generate}

\item{n}{Number of observations per dataset.}

\item{true_betas}{A vector of true values of regression coefficients.
Should include the intercept as the first item, beta1 as the second,
etc.}

\item{select_method}{If leaps, run leaps. Otherwise, if step, run
step-wise selection}

\item{direction}{Only valid of \code{method = "step"}. Run step-wise
selection using \code{step} in the specified direction}

\item{split_prop}{If non-positive, run naive method, where the
confidence #' intervals are calculated on the same data used to
select the variables. Otherwise, if 0 < \code{split_prop} < 1,
randomly subset \code{split_prop} of the data to be in the training
set, and the other part to be in the test set. Run variable selection
on the training set, and report the final confidence interval as
calculated by fitting the selected model on the test set.}

\item{gen_dist}{A function specifiying how to generate the
covariates. Should take \code{n} as its first argument, corresponding
to the number of observations, and \code{p} as the second,
corresponding to the number of dimensions per observation (including
the intercept, so each X_i has p-1 dimensions).}

\item{error_dist}{A function specifying how to generate the errors.
Should take \code{n} as the first argument, specifying the number of
observations}

\item{...}{Additional arguments to gen_dist}
}
\value{
A dataframe, with each row corresponding to one coefficient
that has been selected in one iteration, and columns:

\item{\code{lb}}{Lower bound of the confidence interval}
\item{\code{ub}}{Upper bound of the confidence interval}
\item{\code{coef_name}}{Name of the coefficient}
\item{\code{sim_num}}{Which iteration of the simulation is this
coefficient from?}
}
\description{
Run one replicate of the simulation. Exported to the user.
}

