% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-lib.R
\name{train_test_split}
\alias{train_test_split}
\title{Train-test split}
\usage{
train_test_split(xs, ys, split_prop = 0.5)
}
\arguments{
\item{xs}{Covariates}

\item{ys}{Responses}

\item{split_prop}{Proportion in the training set. Must be between 0
and 1, non-inclusive}
}
\value{
A list with elements
\item{\code{xs_train}}{Training set of covariates}
\item{\code{xs_test}}{Test set of covariates}
\item{\code{ys_train}}{Training set of responses}
\item{\code{ys_test}}{Test set of responses}
}
\description{
Split the data randomly into a training and test set.
NOT exported to the user
}

